+++ src/components/App/index.js
@@ -1,27 +1,14 @@
 import React from 'react';
-import {ShopProvider} from '../../ShopContext.js'
-import {maxBy, minBy} from 'csssr-school-utils/lib/';
+import {ShopProvider, getInitialState} from '../../ShopContext.js'
 import ProductsPage from '../ProductsPage';
 
-const DEFAULT_DISCOUNT = 0;
 
-function getCategory() {
-	const urlParams = new URLSearchParams(window.location.search);
-	return urlParams.get('category');
-}
 
 export default class App extends React.Component {
 	constructor(props) {
 		super(props);
 		
-		this.state = {
-			products: this.props.products,
-			categoriesList: this.props.categoriesList,
-			minPrice: minBy(x => x.price, this.props.products).price,
-			maxPrice: maxBy(x => x.price, this.props.products).price,
-			minDiscount: DEFAULT_DISCOUNT,
-			selectedCategory: getCategory()
-		};
+		this.state = getInitialState();
 	};
 
 	handleChangeMinPrice = (number) => {
@@ -46,7 +33,7 @@ export default class App extends React.Component {
 					selectedCategory: this.state.selectedCategory,
 					
 					categoriesList: this.state.categoriesList,
-					productsData: this.props.products,
+					productsData: this.state.productsData,
 					
 					handleChangeMinPrice: this.handleChangeMinPrice,
 					handleChangeMaxPrice: this.handleChangeMaxPrice,
