+++ src/index.js
@@ -2,20 +2,7 @@ import React from 'react';
 import ReactDOM from 'react-dom';
 import App from './components/App';
 
-import products from './products.json';
-
-const categoriesList = [
-	{
-		id: 0,
-		name: 'Smartphones'
-	},
-	{
-		id: 1,
-		name: 'Accessories'
-	}
-];
-
 ReactDOM.render(
-	<App products={products} categoriesList={categoriesList} />, 
+	<App />, 
 	document.getElementById('root')
 );
